:refactor: Mudança de nome das classes, para uma melhor organização do projeto. / Mudança do Path da ZeldaResult de controller para model.
